requirements.txt
google-generativeai
python-dotenv
requests
beautifulsoup4
jinja2
markdown
